:sparkles: feat: implement useCadastroUsuario hook for user registration logic and refactor Cadastro component to utilize it
